Move ChainIdManager to own module
Move ChainIdManager to its own module, given that it will be shared with
the GenesisController (when it is ported to rust) and to reduce the
number of separate items in chain.rs.
